move to `src`, fix by `pep8`
